pyproject.toml
[tool.poetry]
name = "friendbot"
version = "0.1.0"
description = ""
authors = ["David <161997989+lintyourcode@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.37.1"
python-dotenv = "^1.0.0"
discord = "^2.3.2"
litellm = "^1.42.5"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
